+++ R/formatters.R
@@ -3,9 +3,15 @@
 #' @inheritParams log_level
 #' @return character vector
 #' @export
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_sprintf()], [formatter_glue()], [formatter_glue_safe()], [formatter_glue_or_sprintf()], [formatter_logging()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
-formatter_paste <- structure(function(..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-    eval(paste(...), envir = .topenv)
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_sprintf()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_glue_or_sprintf()],
+#'     [formatter_logging()], [formatter_json()], [formatter_pander()]
+#'     and [skip_formatter()] for marking a string not to apply the
+#'     formatter on it.
+formatter_paste <- structure(function(...,
+                                      .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  paste(...)
 }, generator = quote(formatter_paste()))
 
 
@@ -15,9 +21,15 @@ formatter_paste <- structure(function(..., .logcall = sys.call(), .topcall = sys
 #' @inheritParams log_level
 #' @return character vector
 #' @export
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_paste()], [formatter_glue()], [formatter_glue_safe()], [formatter_glue_or_sprintf()], [formatter_logging()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
-formatter_sprintf <- structure(function(fmt, ..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-    eval(sprintf(fmt, ...), envir = .topenv)
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_paste()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_glue_or_sprintf()],
+#'     [formatter_logging()], [formatter_json()], [formatter_pander()]
+#'     and [skip_formatter()] for marking a string not to apply the
+#'     formatter on it.
+formatter_sprintf <- structure(function(fmt, ...,
+                                        .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  sprintf(fmt, ...)
 }, generator = quote(formatter_sprintf()))
 
 
@@ -26,27 +38,35 @@ formatter_sprintf <- structure(function(fmt, ..., .logcall = sys.call(), .topcal
 #' @inheritParams log_level
 #' @return character vector
 #' @export
-#' @note Although this is the default log message formatter function, but when \pkg{glue} is not installed, [formatter_sprintf()] will be used as a fallback.
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_paste()], [formatter_sprintf()], [formatter_glue_or_sprintf()], [formatter_glue_safe()], [formatter_logging()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
+#' @note Although this is the default log message formatter function,
+#'     but when \pkg{glue} is not installed, [formatter_sprintf()]
+#'     will be used as a fallback.
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_paste()], [formatter_sprintf()],
+#'     [formatter_glue_or_sprintf()], [formatter_glue_safe()],
+#'     [formatter_logging()], [formatter_json()], [formatter_pander()]
+#'     and [skip_formatter()] for marking a string not to apply the
+#'     formatter on it.
 #' @importFrom utils str
-formatter_glue <- structure(function(..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-    fail_on_missing_package('glue')
-    
-    withCallingHandlers(
-        glue::glue(..., .envir = .topenv),
-        error = function(e) {
-            args <- paste0(capture.output(str(...)), collapse = '\n')
+formatter_glue <- structure(function(...,
+                                     .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  fail_on_missing_package("glue")
 
-            stop(paste0(
-                '`glue` failed in `formatter_glue` on:\n\n',
-                args,
-                '\n\nRaw error message:\n\n',
-                conditionMessage(e),
-                '\n\nPlease consider using another `log_formatter` or ',
-                '`skip_formatter` on strings with curly braces.'
-            ))
-        }
-    )
+  withCallingHandlers(
+    glue::glue(..., .envir = .topenv),
+    error = function(e) {
+      args <- paste0(capture.output(str(...)), collapse = "\n")
+
+      stop(paste0(
+        "`glue` failed in `formatter_glue` on:\n\n",
+        args,
+        "\n\nRaw error message:\n\n",
+        conditionMessage(e),
+        "\n\nPlease consider using another `log_formatter` or ",
+        "`skip_formatter` on strings with curly braces."
+      ))
+    }
+  )
 }, generator = quote(formatter_glue()))
 
 
@@ -55,93 +75,113 @@ formatter_glue <- structure(function(..., .logcall = sys.call(), .topcall = sys.
 #' @inheritParams log_level
 #' @return character vector
 #' @export
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_glue()], [formatter_paste()], [formatter_sprintf()], [formatter_glue()], [formatter_glue_or_sprintf()], [formatter_logging()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_glue()], [formatter_paste()], [formatter_sprintf()],
+#'     [formatter_glue()], [formatter_glue_or_sprintf()],
+#'     [formatter_logging()], [formatter_json()], [formatter_pander()]
+#'     and [skip_formatter()] for marking a string not to apply the
+#'     formatter on it.
 #' @importFrom utils str
-formatter_glue_safe <- structure(function(..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-    fail_on_missing_package('glue')
-    as.character(
-        tryCatch(
-            glue::glue_safe(..., .envir = .topenv),
-            error = function(e) {
-                stop(paste(
-                    '`glue_safe` failed in `formatter_glue_safe` on:\n\n',
-                    capture.output(str(...)),
-                    '\n\nRaw error message:\n\n',
-                    e$message,
-                    '\n\nPlease consider using another `log_formatter` or',
-                    '`skip_formatter` on strings with curly braces.'))
-            }))
+formatter_glue_safe <- structure(function(...,
+                                          .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  fail_on_missing_package("glue")
+  as.character(
+    tryCatch(
+      glue::glue_safe(..., .envir = .topenv),
+      error = function(e) {
+        stop(paste(
+          "`glue_safe` failed in `formatter_glue_safe` on:\n\n",
+          capture.output(str(...)),
+          "\n\nRaw error message:\n\n",
+          e$message,
+          "\n\nPlease consider using another `log_formatter` or",
+          "`skip_formatter` on strings with curly braces."
+        ))
+      }
+    )
+  )
 }, generator = quote(formatter_glue_safe()))
 
 
 #' Apply `glue` and `sprintf`
 #'
-#' The best of both words: using both formatter functions in your log messages, which can be useful eg if you are migrating from `sprintf` formatted log messages to `glue` or similar.
+#' The best of both words: using both formatter functions in your log
+#' messages, which can be useful eg if you are migrating from
+#' `sprintf` formatted log messages to `glue` or similar.
 #'
-#' Note that this function tries to be smart when passing arguments to `glue` and `sprintf`, but might fail with some edge cases, and returns an unformatted string.
-#' @param msg passed to `sprintf` as `fmt` or handled as part of `...` in `glue`
+#' Note that this function tries to be smart when passing arguments to
+#' `glue` and `sprintf`, but might fail with some edge cases, and
+#' returns an unformatted string.
+#' @param msg passed to `sprintf` as `fmt` or handled as part of `...`
+#'     in `glue`
 #' @param ... passed to `glue` for the text interpolation
 #' @inheritParams log_level
 #' @return character vector
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_paste()], [formatter_sprintf()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_logging()],
+#'     [formatter_json()], [formatter_pander()] and [skip_formatter()]
+#'     for marking a string not to apply the formatter on it.
 #' @export
 #' @examples \dontrun{
 #' formatter_glue_or_sprintf("{a} + {b} = %s", a = 2, b = 3, 5)
-#' formatter_glue_or_sprintf("{pi} * {2} = %s", pi*2)
+#' formatter_glue_or_sprintf("{pi} * {2} = %s", pi * 2)
 #' formatter_glue_or_sprintf("{pi} * {2} = {pi*2}")
 #'
 #' formatter_glue_or_sprintf("Hi ", "{c('foo', 'bar')}, did you know that 2*4={2*4}")
 #' formatter_glue_or_sprintf("Hi {c('foo', 'bar')}, did you know that 2*4={2*4}")
-#' formatter_glue_or_sprintf("Hi {c('foo', 'bar')}, did you know that 2*4=%s", 2*4)
-#' formatter_glue_or_sprintf("Hi %s, did you know that 2*4={2*4}", c('foo', 'bar'))
-#' formatter_glue_or_sprintf("Hi %s, did you know that 2*4=%s", c('foo', 'bar'), 2*4)
+#' formatter_glue_or_sprintf("Hi {c('foo', 'bar')}, did you know that 2*4=%s", 2 * 4)
+#' formatter_glue_or_sprintf("Hi %s, did you know that 2*4={2*4}", c("foo", "bar"))
+#' formatter_glue_or_sprintf("Hi %s, did you know that 2*4=%s", c("foo", "bar"), 2 * 4)
 #' }
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_paste()], [formatter_sprintf()], [formatter_glue()], [formatter_glue_safe()], [formatter_logging()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
-formatter_glue_or_sprintf <- structure(function(msg, ..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+formatter_glue_or_sprintf <- structure(function(msg, ...,
+                                                .logcall = sys.call(), .topcall = sys.call(-1),
+                                                .topenv = parent.frame()) {
+  params <- list(...)
 
-    params <- list(...)
+  ## params without a name are potential sprintf params
+  sprintfparams <- which(names(params) == "")
+  if (length(params) > 0 & length(sprintfparams) == 0) {
+    sprintfparams <- seq_along(params)
+  }
+  if (is.null(msg) || length(msg) == 0) {
+    msg <- ""
+  }
 
-    ## params without a name are potential sprintf params
-    sprintfparams <- which(names(params) == '')
-    if (length(params) > 0 & length(sprintfparams) == 0) {
-        sprintfparams <- seq_along(params)
-    }
-    if (is.null(msg) || length(msg) == 0) {
-	    msg <- ''
-    }
+  ## early return
+  if (is.null(msg) || length(msg) == 0) {
+    return("")
+  }
 
-    ## early return
-    if (is.null(msg) || length(msg) == 0) {
-        return('')
-    }
+  ## but some unnamed params might belong to glue actually, so
+  ## let's look for the max number of first unnamed params sprintf expects
+  sprintftags <- regmatches(msg, gregexpr("%[0-9.+0]*[aAdifeEgGosxX]", msg))[[1]]
+  sprintfparams <- sprintfparams[seq_len(min(length(sprintftags), length(sprintfparams)))]
 
-    ## but some unnamed params might belong to glue actually, so
-    ## let's look for the max number of first unnamed params sprintf expects
-    sprintftags <- regmatches(msg, gregexpr('%[0-9.+0]*[aAdifeEgGosxX]', msg))[[1]]
-    sprintfparams <- sprintfparams[seq_len(min(length(sprintftags), length(sprintfparams)))]
+  ## get the actual params instead of indexes
+  glueparams <- params[setdiff(seq_along(params), sprintfparams)]
+  sprintfparams <- params[sprintfparams]
 
-    ## get the actual params instead of indexes
-    glueparams    <- params[setdiff(seq_along(params), sprintfparams)]
-    sprintfparams <- params[sprintfparams]
-
-    ## first try to apply sprintf
-    if (length(sprintfparams) > 0) {
-        sprintfparams[vapply(sprintfparams, is.null, logical(1))] <- 'NULL'
-        msg <- tryCatch(
-            do.call(sprintf, c(msg, sprintfparams), envir = .topenv),
-            error = function(e) msg)
-    }
-
-    ## then try to glue
-    fail_on_missing_package('glue')
+  ## first try to apply sprintf
+  if (length(sprintfparams) > 0) {
+    sprintfparams[vapply(sprintfparams, is.null, logical(1))] <- "NULL"
     msg <- tryCatch(
-        as.character(sapply(msg, function(msg) {
-            do.call(glue::glue, c(msg, glueparams), envir = .topenv)
-        }, USE.NAMES = FALSE)),
-        error = function(e) msg)
+      do.call(sprintf, c(msg, sprintfparams), envir = .topenv),
+      error = function(e) msg
+    )
+  }
 
-    ## return
-    msg
+  ## then try to glue
+  fail_on_missing_package("glue")
+  msg <- tryCatch(
+    as.character(sapply(msg, function(msg) {
+      do.call(glue::glue, c(msg, glueparams), envir = .topenv)
+    }, USE.NAMES = FALSE)),
+    error = function(e) msg
+  )
 
+  ## return
+  msg
 }, generator = quote(formatter_glue_or_sprintf()))
 
 
@@ -151,67 +191,89 @@ formatter_glue_or_sprintf <- structure(function(msg, ..., .logcall = sys.call(),
 #' @return character vector
 #' @export
 #' @note This functionality depends on the \pkg{jsonlite} package.
+#' @seealso This is a [log_formatter()] potentially to be used with
+#'     [layout_json_parser()], for alternatives, see
+#'     [formatter_paste()], [formatter_sprintf()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_glue_or_sprintf()],
+#'     [formatter_logging()], [formatter_pander()] and
+#'     [skip_formatter()] for marking a string not to apply the
+#'     formatter on it.
 #' @examples \dontrun{
 #' log_formatter(formatter_json)
 #' log_layout(layout_json_parser())
 #' log_info(everything = 42)
 #' log_info(mtcars = mtcars, species = iris$Species)
 #' }
-#' @seealso This is a [log_formatter()] potentially to be used with [layout_json_parser()], for alternatives, see [formatter_paste()], [formatter_sprintf()], [formatter_glue()], [formatter_glue_safe()], [formatter_glue_or_sprintf()], [formatter_logging()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
-formatter_json <- structure(function(..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-    fail_on_missing_package('jsonlite')
-    eval(as.character(jsonlite::toJSON(list(...), auto_unbox = TRUE)), envir = .topenv)
+formatter_json <- structure(function(...,
+                                     .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  fail_on_missing_package("jsonlite")
+  as.character(jsonlite::toJSON(list(...), auto_unbox = TRUE))
 }, generator = quote(formatter_json()))
 
 
 #' Skip the formatter function
 #'
-#' Adds the `skip_formatter` attribute to an object so that logger will skip calling the formatter function(s). This is useful if you want to preprocess the log message with a custom function instead of the active formatter function(s). Note that the `message` should be a string, and `skip_formatter` should be the only input for the logging function to make this work.
-#' @param message character vector directly passed to the appender function in [logger()]
+#' Adds the `skip_formatter` attribute to an object so that logger
+#' will skip calling the formatter function(s). This is useful if you
+#' want to preprocess the log message with a custom function instead
+#' of the active formatter function(s). Note that the `message` should
+#' be a string, and `skip_formatter` should be the only input for the
+#' logging function to make this work.
+#' @param message character vector directly passed to the appender
+#'     function in [logger()]
 #' @param ... should be never set
-#' @return character vector  with `skip_formatter` attribute set to `TRUE`
+#' @return character vector with `skip_formatter` attribute set to
+#'     `TRUE`
 #' @export
 skip_formatter <- function(message, ...) {
-    if (!inherits(message, 'character')) {
-        stop('Cannot skip the formatter function if the log message is not already formatter to a character vector')
-    }
-    if (length(list(...)) > 0) {
-        stop('Cannot skip the formatter function if further arguments are passed besides the actual log message(s)')
-    }
-    structure(message, skip_formatter = TRUE)
+  if (!inherits(message, "character")) {
+    stop("Cannot skip the formatter function if the log message is not already formatter to a character vector")
+  }
+  if (length(list(...)) > 0) {
+    stop("Cannot skip the formatter function if further arguments are passed besides the actual log message(s)")
+  }
+  structure(message, skip_formatter = TRUE)
 }
 
 
 #' Mimic the default formatter used in the \pkg{logging} package
 #'
-#' The \pkg{logging} package uses a formatter that behaves differently when the input is a string or other R object. If the first argument is a string, then [sprintf()] is being called -- otherwise it does something like [log_eval()] and logs the R expression(s) and the result(s) as well.
-#' @examples \dontrun{
-#' log_formatter(formatter_logging)
-#' log_info('42')
-#' log_info(42)
-#' log_info(4+2)
-#' log_info('foo %s', 'bar')
-#' log_info('vector %s', 1:3)
-#' log_info(12, 1+1, 2 * 2)
-#' }
-#' @param ... string and further params passed to `sprintf` or R expressions to be evaluated
+#' The \pkg{logging} package uses a formatter that behaves differently
+#' when the input is a string or other R object. If the first argument
+#' is a string, then [sprintf()] is being called -- otherwise it does
+#' something like [log_eval()] and logs the R expression(s) and the
+#' result(s) as well.
+#' @param ... string and further params passed to `sprintf` or R
+#'     expressions to be evaluated
 #' @inheritParams log_level
 #' @return character vector
 #' @export
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_paste()], [formatter_glue()], [formatter_glue_safe()], [formatter_glue_or_sprintf()], [formatter_json()], [formatter_pander()] and [skip_formatter()] for marking a string not to apply the formatter on it.
-formatter_logging <- structure(function(..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_paste()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_glue_or_sprintf()],
+#'     [formatter_json()], [formatter_pander()] and [skip_formatter()]
+#'     for marking a string not to apply the formatter on it.
+#' @examples \dontrun{
+#' log_formatter(formatter_logging)
+#' log_info("42")
+#' log_info(42)
+#' log_info(4 + 2)
+#' log_info("foo %s", "bar")
+#' log_info("vector %s", 1:3)
+#' log_info(12, 1 + 1, 2 * 2)
+#' }
+formatter_logging <- structure(function(...,
+                                        .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  params <- list(...)
+  .logcall <- substitute(.logcall)
 
-    params <- list(...)
-    .logcall <- substitute(.logcall)
+  if (is.character(params[[1]])) {
+    return(do.call(sprintf, params, envir = .topenv))
+  }
 
-    if (is.character(params[[1]])) {
-        return(do.call(sprintf, params, envir = .topenv))
-    }
-
-    sapply(1:length(params), function(i) {
-        paste(deparse(as.list(.logcall)[-1][[i]]), params[[i]], sep = ': ')
-    })
-
+  sapply(seq_along(params), function(i) {
+    paste(deparse(as.list(.logcall)[-1][[i]]), params[[i]], sep = ": ")
+  })
 }, generator = quote(formatter_logging()))
 
 
@@ -220,21 +282,23 @@ formatter_logging <- structure(function(..., .logcall = sys.call(), .topcall = s
 #' @param ... optional parameters passed to `pander`
 #' @inheritParams log_level
 #' @return character vector
-#' @examples \dontrun{
-#' log_formatter(formatter_pander)
-#' log_info('42')
-#' log_info(42)
-#' log_info(4+2)
-#' log_info(head(iris))
-#' log_info(head(iris), style = 'simple')
-#' log_info(lm(hp ~ wt, mtcars))
-#' }
 #' @note This functionality depends on the \pkg{pander} package.
 #' @export
-#' @seealso This is a [log_formatter()], for alternatives, see [formatter_paste()], [formatter_sprintf()], [formatter_glue()], [formatter_glue_safe()], [formatter_glue_or_sprintf()], [formatter_logging()]
-formatter_pander <- structure(function(x, ..., .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
-
-    fail_on_missing_package('pander')
-    eval(pander::pander_return(x, ...), envir = .topenv)
-
+#' @seealso This is a [log_formatter()], for alternatives, see
+#'     [formatter_paste()], [formatter_sprintf()], [formatter_glue()],
+#'     [formatter_glue_safe()], [formatter_glue_or_sprintf()],
+#'     [formatter_logging()]
+#' @examples \dontrun{
+#' log_formatter(formatter_pander)
+#' log_info("42")
+#' log_info(42)
+#' log_info(4 + 2)
+#' log_info(head(iris))
+#' log_info(head(iris), style = "simple")
+#' log_info(lm(hp ~ wt, mtcars))
+#' }
+formatter_pander <- structure(function(x, ...,
+                                       .logcall = sys.call(), .topcall = sys.call(-1), .topenv = parent.frame()) {
+  fail_on_missing_package("pander")
+  pander::pander_return(x, ...)
 }, generator = quote(formatter_pander()))